CaD LW№6 & LW№7: fixed the bug with printing of symbols in LW№6 and started using the C PL in LW№7 for the best performance.

+++ The 2nd course/Computers and devices/laboratory work #6/Device.cpp
@@ -79,7 +79,7 @@ string Device::getInterfacesInfo() const {
     const libusb_interface_descriptor* interdesc;
     stringstream interfacesInfo;
     interfacesInfo << right << setw(30) << "Number of interfaces: "
-            << config_->bNumInterfaces << endl;
+            << (int)config_->bNumInterfaces << endl;
     for(int i = 0; i < config_->bNumInterfaces; i++){
         inter = &config_->interface[i];
         interfacesInfo << right << setw(15) << i + 1 << ") "
